refactor(scs): extract modules from monolithic main.rs
Split the 1722-line single-file SCS crate into focused modules:

- auth.rs: repository access and existence checks
- port_forward.rs: port-forward session handling and target resolution
- git/mod.rs: CommandHandler struct, ref collection, ref advertisement
- git/upload_pack.rs: git-upload-pack protocol implementation
- git/receive_pack.rs: git-receive-pack with packfile parsing and delta resolution

main.rs now contains only the CLI, SSH server setup, and session routing.
No behavioral changes; purely structural.
